fix: move about section from components to parts dir
